poc-adk/requirements.txt
google-adk==1.5.0